requirements.txt
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
smtplib-wrapper==0.3.2
pywhatkit==5.4
twilio==8.9.1
schedule==1.2.0
pyodbc==4.0.39